requirements.txt
blinker==1.6.2
click==8.1.7
dateutils==0.6.12
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
Werkzeug==2.3.7
